+++ example/lib/forms_main.dart
@@ -1,0 +1,6 @@
+import 'package:flutter/widgets.dart';
+import 'package:open_ui_kit/open_ui_kit.dart';
+
+import 'form_workflow.dart';
+
+void main() => runApp(const UiApp(home: FormWorkflowExample()));